inevitable/core-identity/requirements.txt
# platform-forge-complete Requirements
# Generated by Platform Forge

# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
# Authentication
pyotp==2.9.0
qrcode==7.4.2

# MFA Email/SMS
twilio==8.10.0
email-validator==2.1.0
# SAML/SSO Dependencies
lxml==4.9.3
xmltodict==0.13.0
requests==2.31.0
# xmlsec==1.3.13  # Temporarily disabled - compilation issues on ARM64
signxml==3.2.2
# python3-saml==1.16.0  # Depends on xmlsec
# Billing
stripe==7.8.0
# Observability
prometheus-client==0.19.0
structlog==23.2.0
# Multi-tenancy
redis==5.0.1
# Security Dependencies
networkx==3.2.1
# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
ruff==0.1.7

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7
jinja2==3.1.2
faker==22.0.0
numpy==1.26.2
argon2-cffi==23.1.0
beautifulsoup4==4.12.2
